+++ RELEASENOTES.md
@@ -2,6 +2,22 @@
 
 ## What is New in CosmosDB Unreleased
 
+February 27, 2018
+
+- Fixed bug in `Get-CosmosDbDocument` when looking up a document in
+  a partitioned collection by adding a `PartitionKey` parameter.
+- Added `Upsert` parameter to `New-CosmosDbDocument` to enable updating
+  a document if it exists.
+- Fixed bug in `New-CosmosDbDocument` when adding document to
+  a partitioned collection but no partition key is specified - See
+  [Issue #48](https://github.com/PlagueHO/CosmosDB/issues/48).
+- Fixed bug in `Set-CosmosDbDocument` when updating a document in
+  a partitioned collection.
+- Fixed bug in `Remove-CosmosDbDocument` when deleting a document in
+  a partitioned collection.
+
+## What is New in CosmosDB 2.0.3.190
+
 February 25, 2018
 
 - Added support for creating custom indexing policies when